+++ library/STMContainers/Bimap.hs
@@ -1,8 +1,6 @@
 module STMContainers.Bimap
 (
   Bimap,
-  Association,
-  Key,
   new,
   newIO,
   insert1,
@@ -36,14 +34,6 @@ data Bimap a b =
   deriving (Typeable)
 
 -- |
--- A constraint for associations.
-type Association a b = (Key a, Key b)
-
--- |
--- A constraint for keys.
-type Key k = Map.Key k
-
--- |
 -- Construct a new bimap.
 {-# INLINABLE new #-}
 new :: STM (Bimap a b)
@@ -73,20 +63,20 @@ size = Map.size . m1
 -- |
 -- Look up a right value by a left value.
 {-# INLINABLE lookup1 #-}
-lookup1 :: (Association a b) => a -> Bimap a b -> STM (Maybe b)
+lookup1 :: (Eq a, Hashable a) => a -> Bimap a b -> STM (Maybe b)
 lookup1 k = Map.lookup k . m1
 
 -- |
 -- Look up a left value by a right value.
 {-# INLINABLE lookup2 #-}
-lookup2 :: (Association a b) => b -> Bimap a b -> STM (Maybe a)
+lookup2 :: (Eq b, Hashable b) => b -> Bimap a b -> STM (Maybe a)
 lookup2 k = Map.lookup k . m2
 
 -- |
 -- Insert an association by a left value.
 {-# INLINABLE insert1 #-}
-insert1 :: (Association a b) => b -> a -> Bimap a b -> STM ()
-insert1 b a (Bimap m1 m2) = 
+insert1 :: (Eq a, Eq b, Hashable a, Hashable b) => b -> a -> Bimap a b -> STM ()
+insert1 b a (Bimap m1 m2) =
   do
     Map.insert b a m1
     Map.insert a b m2
@@ -94,13 +84,13 @@ insert1 b a (Bimap m1 m2) =
 -- |
 -- Insert an association by a right value.
 {-# INLINABLE insert2 #-}
-insert2 :: (Association a b) => a -> b -> Bimap a b -> STM ()
+insert2 :: (Eq a, Eq b, Hashable a, Hashable b) => a -> b -> Bimap a b -> STM ()
 insert2 b a (Bimap m1 m2) = (inline insert1) b a (Bimap m2 m1)
 
 -- |
 -- Delete an association by a left value.
 {-# INLINABLE delete1 #-}
-delete1 :: (Association a b) => a -> Bimap a b -> STM ()
+delete1 :: (Eq a, Eq b, Hashable a, Hashable b) => a -> Bimap a b -> STM ()
 delete1 k (Bimap m1 m2) =
   Map.focus lookupAndDeleteStrategy k m1 >>= 
     mapM_ (\k' -> Map.delete k' m2)
@@ -111,7 +101,7 @@ delete1 k (Bimap m1 m2) =
 -- |
 -- Delete an association by a right value.
 {-# INLINABLE delete2 #-}
-delete2 :: (Association a b) => b -> Bimap a b -> STM ()
+delete2 :: (Eq a, Eq b, Hashable a, Hashable b) => b -> Bimap a b -> STM ()
 delete2 k (Bimap m1 m2) = (inline delete1) k (Bimap m2 m1)
 
 -- |
@@ -131,7 +121,7 @@ deleteAll (Bimap m1 m2) =
 -- E.g., you can look up an item and delete it at the same time,
 -- or update it and return the new value.
 {-# INLINABLE focus1 #-}
-focus1 :: (Association a b) => Focus.StrategyM STM b r -> a -> Bimap a b -> STM r
+focus1 :: (Eq a, Eq b, Hashable a, Hashable b) => Focus.StrategyM STM b r -> a -> Bimap a b -> STM r
 focus1 s a (Bimap m1 m2) =
   do 
     (r, d, mb) <- Map.focus s' a m1
@@ -156,7 +146,7 @@ focus1 s a (Bimap m1 m2) =
 -- E.g., you can look up an item and delete it at the same time,
 -- or update it and return the new value.
 {-# INLINABLE focus2 #-}
-focus2 :: (Association a b) => Focus.StrategyM STM a r -> b -> Bimap a b -> STM r
+focus2 :: (Eq a, Eq b, Hashable a, Hashable b) => Focus.StrategyM STM a r -> b -> Bimap a b -> STM r
 focus2 s b (Bimap m1 m2) = (inline focus1) s b (Bimap m2 m1)
 
 -- |