+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to the
 - Add `STMContainers.Set.member`, `STMContainers.Multimap.member`
 ### Changed
 - Deprecate `STMContainers.Set.lookup`, `STMContainers.Multimap.lookup`
+### Removed
+- Removed constraint `Key`, `Value`, `Association`, and `Element`